lix refactor, impl pr feedback
Summary:

Test Plan:
