+++ .changeset/dry-guests-heal.md
@@ -1,0 +1,5 @@
+---
+"@lix-js/client": minor
+---
+
+lix refactoring
