+++ lix/packages/client/package.json
@@ -36,8 +36,7 @@
 		"octokit": "3.1.2",
 		"pako": "^1.0.11",
 		"pify": "^5.0.0",
-		"sha.js": "^2.4.11",
-		"solid-js": "1.7.11"
+		"sha.js": "^2.4.11"
 	},
 	"license": "Apache-2.0",
 	"devDependencies": {
