+++ lix/packages/client/src/allowedRepos.js
@@ -1,0 +1,11 @@
+export default [
+	"inlang/example",
+	"inlang/ci-test-repo",
+	"opral/monorepo",
+	"inlang/example-test",
+
+	"janfjohannes/inlang-example",
+	"janfjohannes/cal.com",
+
+	"niklasbuchfink/appflowy",
+]
