+++ lix/packages/client/src/git/getCurrentBranch.ts
@@ -1,0 +1,12 @@
+import type { RepoState, RepoContext } from "../openRepository.js"
+import isoGit from "../../vendored/isomorphic-git/index.js"
+
+export async function getCurrentBranch(ctx: RepoContext, state: RepoState) {
+	// TODO: maybe make stateless?
+	return (
+		(await isoGit.currentBranch({
+			fs: state.nodeishFs,
+			dir: ctx.dir,
+		})) || undefined
+	)
+}
