+++ lix/packages/client/src/github/createFork.ts
@@ -1,0 +1,8 @@
+import type { RepoContext } from "../repoContext.js"
+
+export async function createFork(ctx: RepoContext) {
+	return await ctx.githubClient.createFork({
+		owner: ctx.owner,
+		repo: ctx.repoName,
+	})
+}
