+++ lix/packages/client/src/helpers.ts
@@ -169,3 +169,21 @@ export function parseLixUri(uriText: string) {
 		repoName,
 	}
 }
+
+export async function hash(inputStr: string) {
+	let usedCrypto
+	// @ts-ignore
+	if (typeof crypto === "undefined" && typeof process !== "undefined" && process?.versions?.node) {
+		// Use string concatenation to force esbuild to treat module as external and not bundle it
+		const modName = "crypto"
+		usedCrypto = globalThis?.crypto || (await import(`node:${modName}`))
+	} else if (typeof crypto !== "undefined") {
+		usedCrypto = crypto
+	}
+	if (!usedCrypto) {
+		throw new Error("Could not find crypto features in runtime")
+	}
+
+	const idDigest = await usedCrypto.subtle.digest("SHA-256", new TextEncoder().encode(inputStr))
+	return [...new Uint8Array(idDigest)].map((b) => ("00" + b.toString(16)).slice(-2)).join("")
+}
