+++ lix/packages/client/src/index.ts
@@ -1,7 +1,7 @@
-export { type Repository, LixError } from "./api.js"
-export { openRepository, findRepoRoot } from "./openRepository.js"
+export { type Repository, LixError } from "./openRepository.js"
+export { openRepository } from "./openRepository.js"
+export { findRepoRoot } from "./git/findRepoRoot.js"
 export { createNodeishMemoryFs } from "@lix-js/fs"
-export { hash } from "./hash.js"
+export { hash } from "./helpers.js"
 export { mockRepo } from "./mockRepo.js"
-export { listRemotes as _listRemotes } from "../vendored/isomorphic-git/index.js"
 export { getAuthClient } from "./browser-auth.js"
