+++ lix/packages/fs/.eslintrc.json
@@ -1,5 +1,4 @@
 {
-	"extends": [],
 	"rules": {
 		"no-restricted-imports": [
 			"error",
